Move more mercator functionality from wgs84.rs
